+++ server/plugin.go
@@ -120,7 +120,6 @@ func (p *Plugin) OnActivate() error {
 		return mutateSampleLabels(l, target, *p.configuration.HonorTimestamps, []*relabel.Config{})
 	}
 
-	app := p.db.Appender(context.Background())
 	cache := newScrapeCache()
 
 	p.ticker = time.NewTicker(time.Duration(scrapeInterval) * time.Second)
@@ -136,28 +135,29 @@ func (p *Plugin) OnActivate() error {
 		buf := new(bytes.Buffer)
 
 		for range p.ticker.C {
-			content, err := scrapeFn(context.TODO(), "http://"+targetAddress, buf, cfg)
+			// we receive the the appender from tsdb by default
+			app := p.db.Appender(context.Background())
+			buf.Reset()
+
+			ctx, cancel := context.WithTimeout(context.Background(), time.Duration(*cfg.ScrapeTimeoutSeconds)*time.Second)
+			content, err := scrapeFn(ctx, "http://"+targetAddress+"/metrics", buf, cfg)
 			if err != nil {
-				_ = level.Error(logger).Log("msg", "scrape failed", "err", err)
+				cancel()
+				level.Error(logger).Log("msg", "scrape failed", "err", err)
 				continue
 			}
+			cancel()
 
 			err = appendFn(app, logger, cache, buf.Bytes(), content, time.Now().Round(0), cfg)
 			if err != nil {
-				_ = app.Rollback()
-				_ = level.Error(logger).Log("msg", "append failed", "err", err)
+				level.Error(logger).Log("msg", "append failed", "err", err)
 				continue
 			}
 
 			err = app.Commit()
 			if err != nil {
-				_ = level.Error(logger).Log("msg", "scrape commit failed", "err", err)
-				continue
+				level.Error(logger).Log("msg", "scrape commit failed", "err", err)
 			}
-
-			// we receive the the appender from tsdb by default
-			app = p.db.Appender(context.Background())
-			buf.Reset()
 		}
 	}()
